+++ nix/modules/configuration/profiles.nix
@@ -7,34 +7,6 @@ let
 
   profilesSubmodule = types.submodule (
     { name, config, ... }:
-    let
-      data = ''
-        [settings]
-        ${lib.strings.concatMapAttrsStringSep "\n" (
-          name: value: "${name}=${value}"
-        ) config.settings}
-
-        [buildenv]
-        ${lib.concatMapStringsSep "\n" (
-          x: "${x.name}${x.op}${x.value}"
-        ) config.buildEnv}
-
-        [runenv]
-        ${lib.concatMapStringsSep "\n" (
-          x: "${x.name}${x.op}${x.value}"
-        ) config.runEnv}
-
-        [conf]
-        ${lib.strings.concatMapAttrsStringSep "\n" (
-          name: value: "${name}=${value}"
-        ) config.conf}
-
-        [platform_tool_requires]
-        ${lib.strings.concatMapAttrsStringSep "\n" (
-          name: value: "${name}/${value}"
-        ) config.platformToolRequires}
-      '';
-    in
     {
       options = {
         settings = mkOption {
@@ -62,8 +34,14 @@ let
         };
 
         platformToolRequires = mkOption {
-          type = types.attrsOf types.str;
-          description = ''Profile platform tool requires.'';
+          type = types.lazyAttrsOf (types.nullOr types.str);
+          description = ''
+            Profile [platform_tool_requires] section properties.
+
+            These properties are merged with the conan-flake defaults defined
+            in the `defaults.profiles.platformToolRequires` option. Set the
+            entry to `null` to remove that default "require".
+          '';
           default = { };
         };
 
@@ -72,33 +50,32 @@ let
           description = ''
             The profile-outputing derivation generated for the configuration.
           '';
-          default = pkgs.writeText "profile" data;
           defaultText = lib.literalExpression ''
             pkgs.writeText "profile" '''
               [settings]
               ''${lib.strings.concatMapAttrsStringSep "\n" (
                 name: value: "''${name}=''${value}"
-              ) config.settings}
+              ) profiles.settings}
 
               [buildenv]
               ''${lib.strings.concatMapStringSep "\n" (
                 x: "''${x.name}''${x.op}''${x.value}"
-              ) config.buildEnv}
+              ) profiles.buildEnv}
 
               [runenv]
               ''${lib.strings.concatMapStringSep "\n" (
                 x: "''${x.name}''${x.op}''${x.value}"
-              ) config.runEnv}
+              ) profiles.runEnv}
 
               [conf]
               ''${lib.strings.concatMapAttrsStringSep "\n" (
                 name: value: "''${name}=''${value}"
-              ) config.conf}
+              ) profiles.conf}
 
               [platform_tool_requires]
               ''${lib.strings.concatMapAttrsStringSep "\n" (
                 name: value: "''${name}/''${value}"
-              ) config.platformToolRequires}
+              ) final.profiles.platformToolRequires}
             '''
           '';
           readOnly = true;
@@ -106,18 +83,91 @@ let
       };
     }
   );
+
+  data = ''
+    [settings]
+    ${lib.strings.concatMapAttrsStringSep "\n" (
+      name: value: "${name}=${value}"
+    ) config.profiles.settings}
+
+    [buildenv]
+    ${lib.concatMapStringsSep "\n" (
+      x: "${x.name}${x.op}${x.value}"
+    ) config.profiles.buildEnv}
+
+    [runenv]
+    ${lib.concatMapStringsSep "\n" (
+      x: "${x.name}${x.op}${x.value}"
+    ) config.profiles.runEnv}
+
+    [conf]
+    ${lib.strings.concatMapAttrsStringSep "\n" (
+      name: value: "${name}=${value}"
+    ) config.profiles.conf}
+
+    [platform_tool_requires]
+    ${lib.strings.concatMapAttrsStringSep "\n" (
+      name: value: "${name}/${value}"
+    ) config.final.profiles.platformToolRequires}
+  '';
+
+  cfg = config.profiles;
 in
 {
-  options.profiles = mkOption {
-    type = profilesSubmodule;
-    description = ''
-      Conan profiles.
-    '';
+  options = {
+    profiles = mkOption {
+      type = profilesSubmodule;
+      description = ''
+        Conan profiles.
+      '';
+    };
+
+    final.profiles.platformToolRequires = mkOption {
+      type = types.lazyAttrsOf (types.nullOr types.str);
+      readOnly = true;
+      description = ''
+        Final configuration of profile [platform_tool_requires] section
+        properties.
+      '';
+    };
   };
+
   config = {
+    profiles = {
+      settings = { }
+        // lib.optionalAttrs (config.arch != null) {
+        "arch" = config.arch;
+      }
+        // lib.optionalAttrs (config.buildType != null) {
+        "build_type" = config.buildType;
+      }
+        // lib.optionalAttrs (config.compiler != null) {
+        "compiler" = config.compiler;
+      }
+        // lib.optionalAttrs (config.compilerCppStd != null) {
+        "compiler.cppstd" = config.compilerCppStd;
+      }
+        // lib.optionalAttrs (config.compilerLibCxx != null) {
+        "compiler.libcxx" = config.compilerLibCxx;
+      }
+        // lib.optionalAttrs (config.compilerVersion != null) {
+        "compiler.version" = config.compilerVersion;
+      }
+        // lib.optionalAttrs (config.os != null) {
+        "os" = config.os;
+      };
+
+      buildEnv = config.buildEnv;
+      runEnv = config.runEnv;
+      conf = config.conf;
+      text = pkgs.writeText "profile" data;
+    };
+
+    final.profiles.platformToolRequires = config.defaults.profiles.platformToolRequires // cfg.platformToolRequires;
+
     outputs = {
       configuration.profile = {
-        package = config.profiles.text;
+        package = cfg.text;
         manifest = "config/profiles/default";
         kind = "configuration";
       };