+++ test/flake-parts-no-defaults/flake.nix
@@ -34,7 +34,7 @@
 
             inherit configLocal conanHome buildType compilerCppStd compilerLibCxx;
 
-            platformToolRequires = {
+            profiles.platformToolRequires = {
               cmake = pkgs.cmake.version;
             };